Compatibility with Python 3.8, moved functions for multiprocessing to util.apply_func
