+++ .readthedocs.yml
@@ -7,6 +7,6 @@ build:
     image: latest
 
 python:
-   version: 3.7
+   version: 3.8
    install:
       - requirements: requirements.txt
