+++ .travis.yml
@@ -7,6 +7,7 @@ dist: xenial
 python:
   - 3.6
   - 3.7
+  - 3.8
 
 env:
   - NUMBA_DISABLE_JIT=1
