+++ docs/about.rst
@@ -8,17 +8,17 @@ About
 Development Team
 ----------------
 
-* Tomas Stolker <tomas.stolker@phys.ethz.ch>
-* Markus Bonse <markus.bonse@stud.tu-darmstadt.de>
+* Tomas Stolker <stolker@strw.leidenuniv.nl>
+* Markus Bonse <mbonse@phys.ethz.ch>
 * Sascha Quanz <sascha.quanz@phys.ethz.ch>
-* Adam Amara <adam.amara@phys.ethz.ch>
+* Adam Amara <adam.amara@port.ac.uk>
 
 .. _attribution:
 
 Attribution
 -----------
 
-If you use PynPoint in your publication then please cite `Stolker et al. (2019) <http://ui.adsabs.harvard.edu/abs/2019A%26A...621A..59S>`_. Please also cite `Amara & Quanz (2012) <http://ui.adsabs.harvard.edu/abs/2012MNRAS.427..948A>`_ as the origin of PynPoint, which focused initially on the use of principal component analysis (PCA) as a PSF subtraction method. In case you use specifically the PCA-based background subtraction module or the wavelet based speckle suppression module, please give credit to `Hunziker et al. (2018) <http://ui.adsabs.harvard.edu/abs/2018A%26A...611A..23H>`_ or `Bonse, Quanz & Amara (2018) <http://ui.adsabs.harvard.edu/abs/2018arXiv180405063B>`_, respectively.
+If you use PynPoint in your publication then please cite `Stolker et al. (2019) <http://ui.adsabs.harvard.edu/abs/2019A%26A...621A..59S>`_. Please also cite `Amara & Quanz (2012) <http://ui.adsabs.harvard.edu/abs/2012MNRAS.427..948A>`_ as the origin of PynPoint, which focused initially on the use of principal component analysis (PCA) as a PSF subtraction method. In case you use specifically the PCA-based background subtraction module or the wavelet based speckle suppression module, please give credit to `Hunziker et al. (2018) <http://ui.adsabs.harvard.edu/abs/2018A%26A...611A..23H>`_ or `Bonse et al. (preprint) <http://ui.adsabs.harvard.edu/abs/2018arXiv180405063B>`_, respectively.
 
 .. _acknowledgements:
