+++ docs/contributing.rst
@@ -9,7 +9,7 @@ We also welcome active help with bug fixing and the development of new functiona
 
 .. |stolker| raw:: html
 
-   <a href="https://people.phys.ethz.ch/~stolkert/" target="_blank">Tomas Stolker</a>
+   <a href="https://home.strw.leidenuniv.nl/~stolker/" target="_blank">Tomas Stolker</a>
 
 .. |issue| raw:: html
