+++ pynpoint/__init__.py
@@ -103,5 +103,5 @@ __author__ = 'Tomas Stolker & Markus Bonse'
 __license__ = 'GPLv3'
 __version__ = '0.8.3'
 __maintainer__ = 'Tomas Stolker'
-__email__ = 'tomas.stolker@phys.ethz.ch'
+__email__ = 'stolker@strw.leidenuniv.nl'
 __status__ = 'Development'
