+++ pynpoint/processing/centering.py
@@ -2,24 +2,22 @@
 Pipeline modules for aligning and centering of the star.
 """
 
-import time
 import math
+import time
 import warnings
 
 from typing import Optional, Tuple, Union
 
 import numpy as np
 
-from astropy.modeling import models, fitting
+from astropy.modeling import fitting, models
 from scipy.ndimage.filters import gaussian_filter
-from scipy.optimize import curve_fit
-from skimage.registration import phase_cross_correlation
-from skimage.transform import rescale
 from typeguard import typechecked
 
 from pynpoint.core.processing import ProcessingModule
 from pynpoint.util.module import memory_frames, progress
-from pynpoint.util.image import crop_image, shift_image, center_pixel
+from pynpoint.util.image import center_pixel, crop_image, shift_image
+from pynpoint.util.apply_func import align_image, fit_2d_function
 
 
 class StarAlignmentModule(ProcessingModule):
@@ -102,49 +100,6 @@ class StarAlignmentModule(ProcessingModule):
             None
         """
 
-        @typechecked
-        def _align_image(image_in: np.ndarray) -> np.ndarray:
-            offset = np.array([0., 0.])
-
-            for i in range(self.m_num_references):
-                if self.m_subframe is None:
-                    tmp_offset, _, _ = phase_cross_correlation(ref_images[i, :, :],
-                                                               image_in,
-                                                               upsample_factor=self.m_accuracy)
-
-                else:
-                    sub_in = crop_image(image_in, None, self.m_subframe)
-                    sub_ref = crop_image(ref_images[i, :, :], None, self.m_subframe)
-
-                    tmp_offset, _, _ = phase_cross_correlation(sub_ref,
-                                                               sub_in,
-                                                               upsample_factor=self.m_accuracy)
-                offset += tmp_offset
-
-            offset /= float(self.m_num_references)
-
-            if self.m_resize is not None:
-                offset *= self.m_resize
-
-                sum_before = np.sum(image_in)
-
-                tmp_image = rescale(image_in,
-                                    (self.m_resize, self.m_resize),
-                                    order=5,
-                                    mode='reflect',
-                                    multichannel=False,
-                                    anti_aliasing=True)
-
-                sum_after = np.sum(tmp_image)
-
-                # Conserve flux because the rescale function normalizes all values to [0:1].
-                tmp_image = tmp_image*(sum_before/sum_after)
-
-            else:
-                tmp_image = image_in
-
-            return shift_image(tmp_image, offset, self.m_interpolation)
-
         if self.m_ref_image_in_port is None:
             random = np.random.choice(self.m_image_in_port.get_shape()[0],
                                       self.m_num_references,
@@ -169,10 +124,17 @@ class StarAlignmentModule(ProcessingModule):
             pixscale = self.m_image_in_port.get_attribute('PIXSCALE')
             self.m_subframe = int(self.m_subframe/pixscale)
 
-        self.apply_function_to_images(_align_image,
+        self.apply_function_to_images(align_image,
                                       self.m_image_in_port,
                                       self.m_image_out_port,
-                                      'Aligning images')
+                                      'Aligning images',
+                                      func_args=(self.m_interpolation,
+                                                 self.m_accuracy,
+                                                 self.m_resize,
+                                                 self.m_num_references,
+                                                 self.m_subframe,
+                                                 ref_images.reshape(-1),
+                                                 ref_images.shape))
 
         self.m_image_out_port.copy_attributes(self.m_image_in_port)
 
@@ -180,7 +142,7 @@ class StarAlignmentModule(ProcessingModule):
             pixscale = self.m_image_in_port.get_attribute('PIXSCALE')
             new_pixscale = pixscale/self.m_resize
             self.m_image_out_port.add_attribute('PIXSCALE', new_pixscale)
-            print(f'New pixel scale [arcsec] = {new_pixscale:.2f}')
+            print(f'New pixel scale (arcsec) = {new_pixscale:.2f}')
 
         history = f'resize = {self.m_resize}'
         self.m_image_out_port.add_history('StarAlignmentModule', history)
@@ -335,229 +297,26 @@ class FitCenterModule(ProcessingModule):
         xx_ap, yy_ap = np.meshgrid(x_ap, y_ap)
         rr_ap = np.sqrt(xx_ap**2+yy_ap**2)
 
-        @typechecked
-        def gaussian_2d(grid: Union[Tuple[np.ndarray, np.ndarray], np.ndarray],
-                        x_center: float,
-                        y_center: float,
-                        fwhm_x: float,
-                        fwhm_y: float,
-                        amp: float,
-                        theta: float,
-                        offset: float) -> np.ndarray:
-            """
-            Function to create a 2D elliptical Gaussian model.
-
-            Parameters
-            ----------
-            grid : tuple(numpy.ndarray, numpy.ndarray), numpy.ndarray
-                A tuple of two 2D arrays with the mesh grid points in x and y
-                direction, or an equivalent 3D numpy array with 2 elements
-                along the first axis.
-            x_center : float
-                Offset of the model center along the x axis (pix).
-            y_center : float
-                Offset of the model center along the y axis (pix).
-            fwhm_x : float
-                Full width at half maximum along the x axis (pix).
-            fwhm_y : float
-                Full width at half maximum along the y axis (pix).
-            amp : float
-                Peak flux.
-            theta : float
-                Rotation angle in counterclockwise direction (rad).
-            offset : float
-                Flux offset.
-
-            Returns
-            -------
-            numpy.ndimage
-                Raveled 2D elliptical Gaussian model.
-            """
-
-            (xx_grid, yy_grid) = grid
-
-            x_diff = xx_grid - x_center
-            y_diff = yy_grid - y_center
-
-            sigma_x = fwhm_x/math.sqrt(8.*math.log(2.))
-            sigma_y = fwhm_y/math.sqrt(8.*math.log(2.))
-
-            a_gauss = 0.5 * ((np.cos(theta)/sigma_x)**2 + (np.sin(theta)/sigma_y)**2)
-            b_gauss = 0.5 * ((np.sin(2.*theta)/sigma_x**2) - (np.sin(2.*theta)/sigma_y**2))
-            c_gauss = 0.5 * ((np.sin(theta)/sigma_x)**2 + (np.cos(theta)/sigma_y)**2)
-
-            gaussian = offset + amp*np.exp(-(a_gauss*x_diff**2 + b_gauss*x_diff*y_diff +
-                                             c_gauss*y_diff**2))
-
-            if self.m_radius:
-                gaussian = gaussian[rr_ap < self.m_radius]
-            else:
-                gaussian = np.ravel(gaussian)
-
-            return gaussian
-
-        @typechecked
-        def moffat_2d(grid: Union[Tuple[np.ndarray, np.ndarray], np.ndarray],
-                      x_center: float,
-                      y_center: float,
-                      fwhm_x: float,
-                      fwhm_y: float,
-                      amp: float,
-                      theta: float,
-                      offset: float,
-                      beta: float) -> np.ndarray:
-            """
-            Function to create a 2D elliptical Moffat model.
-
-            The parametrization used here is equivalent to the one in AsPyLib:
-            http://www.aspylib.com/doc/aspylib_fitting.html#elliptical-moffat-psf
-
-            Parameters
-            ----------
-            grid : tuple(numpy.ndarray, numpy.ndarray), numpy.ndarray
-                A tuple of two 2D arrays with the mesh grid points in x and y
-                direction, or an equivalent 3D numpy array with 2 elements
-                along the first axis.
-            x_center : float
-                Offset of the model center along the x axis (pix).
-            y_center : float
-                Offset of the model center along the y axis (pix).
-            fwhm_x : float
-                Full width at half maximum along the x axis (pix).
-            fwhm_y : float
-                Full width at half maximum along the y axis (pix).
-            amp : float
-                Peak flux.
-            theta : float
-                Rotation angle in counterclockwise direction (rad).
-            offset : float
-                Flux offset.
-            beta : float
-                Power index.
-
-            Returns
-            -------
-            numpy.ndimage
-                Raveled 2D elliptical Moffat model.
-            """
-
-            (xx_grid, yy_grid) = grid
-
-            x_diff = xx_grid - x_center
-            y_diff = yy_grid - y_center
-
-            if 2.**(1./beta)-1. < 0.:
-                alpha_x = np.nan
-                alpha_y = np.nan
-
-            else:
-                alpha_x = 0.5*fwhm_x/np.sqrt(2.**(1./beta)-1.)
-                alpha_y = 0.5*fwhm_y/np.sqrt(2.**(1./beta)-1.)
-
-            if alpha_x == 0. or alpha_y == 0.:
-                a_moffat = np.nan
-                b_moffat = np.nan
-                c_moffat = np.nan
-
-            else:
-                a_moffat = (np.cos(theta)/alpha_x)**2. + (np.sin(theta)/alpha_y)**2.
-                b_moffat = (np.sin(theta)/alpha_x)**2. + (np.cos(theta)/alpha_y)**2.
-                c_moffat = 2.*np.sin(theta)*np.cos(theta)*(1./alpha_x**2. - 1./alpha_y**2.)
-
-            a_term = a_moffat*x_diff**2
-            b_term = b_moffat*y_diff**2
-            c_term = c_moffat*x_diff*y_diff
-
-            moffat = offset + amp / (1.+a_term+b_term+c_term)**beta
-
-            if self.m_radius:
-                moffat = moffat[rr_ap < self.m_radius]
-            else:
-                moffat = np.ravel(moffat)
-
-            return moffat
-
-        @typechecked
-        def _fit_2d_function(image: np.ndarray) -> np.ndarray:
-
-            if self.m_filter_size:
-                image = gaussian_filter(image, self.m_filter_size)
-
-            if self.m_mask_out_port:
-                mask = np.copy(image)
-
-                if self.m_radius:
-                    mask[rr_ap > self.m_radius] = 0.
-
-                self.m_mask_out_port.append(mask, data_dim=3)
-
-            if self.m_sign == 'negative':
-                image = -1.*image + np.abs(np.min(-1.*image))
-
-            if self.m_radius:
-                image = image[rr_ap < self.m_radius]
-            else:
-                image = np.ravel(image)
-
-            if self.m_model == 'gaussian':
-                self.m_model_func = gaussian_2d
-
-            elif self.m_model == 'moffat':
-                self.m_model_func = moffat_2d
-
-            try:
-                popt, pcov = curve_fit(self.m_model_func,
-                                       (xx_grid, yy_grid),
-                                       image,
-                                       p0=self.m_guess,
-                                       sigma=None,
-                                       method='lm')
-
-                perr = np.sqrt(np.diag(pcov))
-
-            except RuntimeError:
-                if self.m_model == 'gaussian':
-                    popt = np.zeros(7)
-                    perr = np.zeros(7)
-
-                elif self.m_model == 'moffat':
-                    popt = np.zeros(8)
-                    perr = np.zeros(8)
-
-                self.m_count += 1
-
-            if self.m_model == 'gaussian':
-
-                best_fit = np.asarray((popt[0], perr[0],
-                                       popt[1], perr[1],
-                                       popt[2]*pixscale, perr[2]*pixscale,
-                                       popt[3]*pixscale, perr[3]*pixscale,
-                                       popt[4], perr[4],
-                                       math.degrees(popt[5]) % 360., math.degrees(perr[5]),
-                                       popt[6], perr[6]))
-
-            elif self.m_model == 'moffat':
-
-                best_fit = np.asarray((popt[0], perr[0],
-                                       popt[1], perr[1],
-                                       popt[2]*pixscale, perr[2]*pixscale,
-                                       popt[3]*pixscale, perr[3]*pixscale,
-                                       popt[4], perr[4],
-                                       math.degrees(popt[5]) % 360., math.degrees(perr[5]),
-                                       popt[6], perr[6],
-                                       popt[7], perr[7]))
-
-            return best_fit
-
         nimages = self.m_image_in_port.get_shape()[0]
         frames = memory_frames(memory, nimages)
 
         if self.m_method == 'full':
 
-            self.apply_function_to_images(_fit_2d_function,
+            self.apply_function_to_images(fit_2d_function,
                                           self.m_image_in_port,
                                           self.m_fit_out_port,
-                                          'Fitting the stellar PSF')
+                                          'Fitting the stellar PSF',
+                                          func_args=(self.m_radius,
+                                                     self.m_sign,
+                                                     self.m_model,
+                                                     self.m_filter_size,
+                                                     self.m_model_func,
+                                                     self.m_guess,
+                                                     self.m_mask_out_port,
+                                                     xx_grid,
+                                                     yy_grid,
+                                                     rr_ap,
+                                                     pixscale))
 
         elif self.m_method == 'mean':
             print('Fitting the stellar PSF...', end='')
@@ -567,16 +326,25 @@ class FitCenterModule(ProcessingModule):
             for i, _ in enumerate(frames[:-1]):
                 im_mean += np.sum(self.m_image_in_port[frames[i]:frames[i+1], ], axis=0)
 
-            best_fit = _fit_2d_function(im_mean/float(nimages))
+            best_fit = fit_2d_function(im_mean/float(nimages),
+                                       self.m_radius,
+                                       self.m_sign,
+                                       self.m_model,
+                                       self.m_filter_size,
+                                       self.m_model_func,
+                                       self.m_guess,
+                                       self.m_mask_out_port,
+                                       xx_grid,
+                                       yy_grid,
+                                       rr_ap,
+                                       pixscale)
+
             best_fit = best_fit[np.newaxis, ...]
             best_fit = np.repeat(best_fit, nimages, axis=0)
 
             self.m_fit_out_port.set_all(best_fit, data_dim=2)
 
             print(' [DONE]')
-
-        if self.m_count > 0:
-            print(f'Fit could not converge on {self.m_count} image(s). [WARNING]')
 
         history = f'model = {self.m_model}'
