+++ pynpoint/processing/frameselection.py
@@ -16,6 +16,7 @@ from typeguard import typechecked
 from skimage.metrics import structural_similarity, mean_squared_error
 
 from pynpoint.core.processing import ProcessingModule
+from pynpoint.util.apply_func import image_stat
 from pynpoint.util.image import crop_image, pixel_distance, center_pixel
 from pynpoint.util.module import progress
 from pynpoint.util.remove import write_selected_data, write_selected_attributes
@@ -618,27 +619,7 @@ class ImageStatisticsModule(ProcessingModule):
             rr_reshape = np.reshape(rr_grid, (rr_grid.shape[0]*rr_grid.shape[1]))
             indices = np.where(rr_reshape <= self.m_position[2])[0]
 
-        @typechecked
-        def _image_stat(image_in: np.ndarray,
-                        indices: Optional[np.ndarray]) -> np.ndarray:
-
-            if indices is None:
-                image_select = np.copy(image_in)
-
-            else:
-                image_reshape = np.reshape(image_in, (image_in.shape[0]*image_in.shape[1]))
-                image_select = image_reshape[indices]
-
-            nmin = np.nanmin(image_select)
-            nmax = np.nanmax(image_select)
-            nsum = np.nansum(image_select)
-            mean = np.nanmean(image_select)
-            median = np.nanmedian(image_select)
-            std = np.nanstd(image_select)
-
-            return np.asarray([nmin, nmax, nsum, mean, median, std])
-
-        self.apply_function_to_images(_image_stat,
+        self.apply_function_to_images(image_stat,
                                       self.m_image_in_port,
                                       self.m_stat_out_port,
                                       'Calculating image statistics',
