+++ pynpoint/processing/psfsubtraction.py
@@ -16,6 +16,7 @@ from sklearn.decomposition import PCA
 from typeguard import typechecked
 
 from pynpoint.core.processing import ProcessingModule
+from pynpoint.util.apply_func import subtract_psf
 from pynpoint.util.module import progress
 from pynpoint.util.multipca import PcaMultiprocessingCapsule
 from pynpoint.util.residuals import combine_residuals
@@ -81,8 +82,8 @@ class PcaPsfSubtractionModule(ProcessingModule):
             Database entry where the derotated, but not collapsed, residuals are stored. The number
             of principal components is was used is appended to the ``res_arr_out_tag``. The
             residuals are not stored if set to None. This parameter is not supported with
-            multiprocessing (i.e. ``CPU`` > 1). For IFS data and if the processing type is either 
-            ADI+SDI or SDI+ADI the residuals can only be calculated if exactly 1 principal component 
+            multiprocessing (i.e. ``CPU`` > 1). For IFS data and if the processing type is either
+            ADI+SDI or SDI+ADI the residuals can only be calculated if exactly 1 principal component
             for each ADI and SDI is given with the pca_numbers parameter.
         basis_out_tag : str, None
             Database entry where the principal components are stored. The data is not stored if set
@@ -190,7 +191,9 @@ class PcaPsfSubtractionModule(ProcessingModule):
                 raise ValueError(f'The post-processing type \'{self.m_processing_type}\' requires '
                                  f'a tuple for with twice a range/list/array as argument for '
                                  f'\'pca_numbers\'.')
-            if res_arr_out_tag is not None and len(self.m_components[0]) + len(self.m_components[1]) != 2:
+
+            if res_arr_out_tag is not None and len(self.m_components[0]) + \
+                    len(self.m_components[1]) != 2:
                 raise ValueError(f'If the post-processing type \'{self.m_processing_type}\' '
                                  'is selected, residuals can only be calculated if no more than '
                                  '1 principal component for ADI and SDI is given.')
@@ -236,7 +239,8 @@ class PcaPsfSubtractionModule(ProcessingModule):
                 res_shape = (len(pca_first), len(wavelength), im_shape[-2], im_shape[-1])
 
             elif self.m_processing_type in ['SDI+ADI', 'ADI+SDI']:
-                res_shape = (len(pca_first), len(pca_secon), len(wavelength), im_shape[-2], im_shape[-1])
+                res_shape = (len(pca_first), len(pca_secon), len(wavelength),
+                             im_shape[-2], im_shape[-1])
 
         else:
             res_shape = (len(self.m_components), im_shape[1], im_shape[2])
@@ -431,8 +435,8 @@ class PcaPsfSubtractionModule(ProcessingModule):
 
             else:
                 history = f'max PC number = {pca_first} / {pca_secon}'
-            
-            # squeeze out_array_res to reduce dimensionallity as the residuals of 
+
+            # squeeze out_array_res to reduce dimensionallity as the residuals of
             # SDI+ADI and ADI+SDI are always of the form (1, 1, ...)
             squeezed = np.squeeze(out_array_res)
 
@@ -544,7 +548,8 @@ class PcaPsfSubtractionModule(ProcessingModule):
             print('Constructing PSF model...', end='')
             self.m_pca.fit(ref_reshape)
 
-            # add mean of reference array as 1st PC and orthogonalize it with respect to the PCA basis
+            # add mean of reference array as 1st PC and orthogonalize it with respect to
+            # the other principal components
             if not self.m_subtract_mean:
                 mean_ref_reshape = mean_ref.reshape((1, mean_ref.shape[0]))
 
@@ -669,8 +674,6 @@ class ClassicalADIModule(ProcessingModule):
         self.m_extra_rot = extra_rot
         self.m_residuals = residuals
 
-        self.m_count = 0
-
     @typechecked
     def run(self) -> None:
         """
@@ -687,37 +690,6 @@ class ClassicalADIModule(ProcessingModule):
             None
         """
 
-        @typechecked
-        def _subtract_psf(image: np.ndarray,
-                          parang_thres: Optional[float],
-                          nref: Optional[int],
-                          reference: Optional[np.ndarray] = None) -> np.ndarray:
-
-            if parang_thres:
-                ang_diff = np.abs(parang[self.m_count]-parang)
-                index_thres = np.where(ang_diff > parang_thres)[0]
-
-                if index_thres.size == 0:
-                    reference = self.m_image_in_port.get_all()
-                    warnings.warn('No images meet the rotation threshold. Creating a reference '
-                                  'PSF from the median of all images instead.')
-
-                else:
-                    if nref:
-                        index_diff = np.abs(self.m_count - index_thres)
-                        index_near = np.argsort(index_diff)[:nref]
-                        index_sort = np.sort(index_thres[index_near])
-                        reference = self.m_image_in_port[index_sort, :, :]
-
-                    else:
-                        reference = self.m_image_in_port[index_thres, :, :]
-
-                reference = np.median(reference, axis=0)
-
-            self.m_count += 1
-
-            return image-reference
-
         parang = -1.*self.m_image_in_port.get_attribute('PARANG') + self.m_extra_rot
 
         if self.m_threshold:
@@ -731,11 +703,14 @@ class ClassicalADIModule(ProcessingModule):
             reference = self.m_image_in_port.get_all()
             reference = np.median(reference, axis=0)
 
-        self.apply_function_to_images(_subtract_psf,
+        self.apply_function_to_images(subtract_psf,
                                       self.m_image_in_port,
                                       self.m_res_out_port,
                                       'Classical ADI',
-                                      func_args=(parang_thres, self.m_nreference, reference))
+                                      func_args=(parang_thres,
+                                                 self.m_nreference,
+                                                 reference,
+                                                 self.m_image_in_port))
 
         self.m_res_in_port = self.add_input_port(self.m_res_out_port._m_tag)
         im_res = self.m_res_in_port.get_all()
