+++ pynpoint/processing/resizing.py
@@ -5,15 +5,16 @@ Pipeline modules for resizing of images.
 import math
 import time
 
-from typing import Union, Tuple
+from typing import Tuple, Union
 
 import numpy as np
 
 from typeguard import typechecked
 
 from pynpoint.core.processing import ProcessingModule
-from pynpoint.util.image import crop_image, scale_image
-from pynpoint.util.module import progress, memory_frames
+from pynpoint.util.apply_func import image_scaling
+from pynpoint.util.image import crop_image
+from pynpoint.util.module import memory_frames, progress
 
 
 class CropImagesModule(ProcessingModule):
@@ -110,7 +111,7 @@ class CropImagesModule(ProcessingModule):
 
             # Update progress bar
             progress(i, len(frames[:-1]), 'Cropping images...', start_time)
-            
+
             # Select images in the current chunk
             if ndim == 3:
                 images = self.m_image_in_port[frames[i]:frames[i+1], ]
@@ -118,7 +119,7 @@ class CropImagesModule(ProcessingModule):
             elif ndim == 4:
                 # Process all wavelengths per exposure at once
                 images = self.m_image_in_port[:, i, ]
-            
+
             # crop images according to input parameters
             images = crop_image(images, self.m_center, self.m_size, copy=False)
 
@@ -208,15 +209,7 @@ class ScaleImagesModule(ProcessingModule):
 
         pixscale = self.m_image_in_port.get_attribute('PIXSCALE')
 
-        @typechecked
-        def _image_scaling(image_in: np.ndarray,
-                           scaling_y: float,
-                           scaling_x: float,
-                           scaling_flux: float) -> np.ndarray:
-
-            return scaling_flux * scale_image(image_in, scaling_y, scaling_x)
-
-        self.apply_function_to_images(_image_scaling,
+        self.apply_function_to_images(image_scaling,
                                       self.m_image_in_port,
                                       self.m_image_out_port,
                                       'Scaling images',
