+++ tests/test_processing/test_limits.py
@@ -133,7 +133,8 @@ class TestLimits:
         with h5py.File(self.test_dir+'PynPoint_database.hdf5', 'a') as hdf_file:
             hdf_file['contrast_limits'] = limits
 
-        url = 'https://people.phys.ethz.ch/~stolkert/pynpoint/model.AMES-Cond-2000.M-0.0.NaCo.Vega'
+        url = 'https://home.strw.leidenuniv.nl/~stolker/pynpoint/' \
+              'model.AMES-Cond-2000.M-0.0.NaCo.Vega'
 
         filename = self.test_dir + 'model.AMES-Cond-2000.M-0.0.NaCo.Vega'
