+++ tests/test_processing/test_psfsubtraction_sdi.py
@@ -47,19 +47,18 @@ class TestPsfSubtractionSdi:
 
         processing_types = ['ADI', 'SDI+ADI', 'ADI+SDI']
 
-        expected = [[-0.16718942968552664, -0.790697125718532, 19.507979777136892, -0.21617058715490922],
-                    [-0.001347198747121658, -0.08621264803633322, 2.3073192270025333, -0.010269745733878437],
-                    [0.009450917836998779, -0.05776205365084376, -0.43506678222476264, 0.0058856438951644455]]
+        expected = [[-0.16718942968552664, -0.790697125718532,
+                     19.507979777136892, -0.21617058715490922],
+                    [-0.001347198747121658, -0.08621264803633322,
+                     2.3073192270025333, -0.010269745733878437],
+                    [0.009450917836998779, -0.05776205365084376,
+                     -0.43506678222476264, 0.0058856438951644455]]
 
         shape_expc = [(2, 3, 21, 21), (2, 2, 3, 21, 21), (1, 1, 3, 21, 21)]
 
         pca_numbers = [range(1, 3), (range(1, 3), range(1, 3)), ([1], [1])]
-        
-        res_arr_tags = [None, None, 'res_arr_single_sdi_ADI+SDI']
 
-        depr_warn = [False, True, True, False,
-                     False, False, False, False,
-                     False, False, False, False]
+        res_arr_tags = [None, None, 'res_arr_single_sdi_ADI+SDI']
 
         for i, p_type in enumerate(processing_types):
 
@@ -78,13 +77,7 @@ class TestPsfSubtractionSdi:
                                              processing_type=p_type)
 
             self.pipeline.add_module(module)
-
-            if depr_warn[i]:
-                with pytest.warns(DeprecationWarning):
-                    self.pipeline.run_module('pca_single_sdi_'+p_type)
-
-            else:
-                self.pipeline.run_module('pca_single_sdi_'+p_type)
+            self.pipeline.run_module('pca_single_sdi_'+p_type)
 
             data = self.pipeline.get_data('res_mean_single_sdi_'+p_type)
             assert np.sum(data) == pytest.approx(expected[i][0], rel=self.limit, abs=0.)
@@ -115,8 +108,10 @@ class TestPsfSubtractionSdi:
 
         pca_numbers = [range(1, 3), (range(1, 3), range(1, 3))]
 
-        expected = [[-0.004159475403024583, 0.02613693149969979, -0.12940723035023394, -0.008432530081399985],
-                    [-0.006580571531064533, -0.08171546066331437, 0.5700432018961117, -0.014527353460544753]]
+        expected = [[-0.004159475403024583, 0.02613693149969979,
+                     -0.12940723035023394, -0.008432530081399985],
+                    [-0.006580571531064533, -0.08171546066331437,
+                     0.5700432018961117, -0.014527353460544753]]
 
         shape_expc = [(2, 3, 21, 21), (2, 2, 3, 21, 21)]
